requirements.txt
fastapi
uvicorn
google-cloud-firestore
httpx
pydantic>=2.0.0
python-dotenv
